# backend/openapi.yaml
openapi: 3.0.0
info:
  title: ReviewTalk Backend API
  version: 1.0.0
  description: |-
    제품 리뷰 기반 AI 챗봇 백엔드 API 명세입니다.
    FastAPI로 구축되었으며, 채팅 기능 및 리뷰 데이터를 활용한 AI 응답 기능을 제공합니다.
    추후 users/{user_id} 는 header 값으로 전달 받아서 인증을 하는 식으로 변경할 예정입니다.
    일단 편의를 위해 users/{user_id} 로 사용합니다.

servers:
  - url: http://localhost:8000
    description: Local Development Server

paths:
  /ping:
    get:
      summary: Health Check
      operationId: ping
      responses:
        "200":
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: pong

  # userId 기반 채팅 목록 관리
  /api/v1/users/{user_id}/chats:
    get:
      summary: 특정 사용자의 채팅 목록 조회
      operationId: get_user_chats
      parameters:
        - name: user_id
          in: path
          required: true
          schema:
            type: string
          description: 사용자 ID
      responses:
        "200":
          description: 성공적인 조회
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ChatListItem"
        "404":
          description: 사용자를 찾을 수 없음
        "500":
          $ref: "#/components/responses/InternalServerError"

    post:
      summary: 특정 사용자의 새 채팅 생성
      operationId: create_user_chat
      parameters:
        - name: user_id
          in: path
          required: true
          schema:
            type: string
          description: 사용자 ID
        - name: url
          in: query
          required: true
          schema:
            type: string
          description: 채팅 생성시 연결될 제품에 대한 URL 
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateChatRequest"
      responses:
        "201":
          description: 채팅 생성 성공
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChatListItem"
        "400":
          description: 잘못된 요청
        "404":
          description: 사용자를 찾을 수 없음
        "500":
          $ref: "#/components/responses/InternalServerError"

  /api/v1/users/{user_id}/chats/{chat_id}:
    # 기존 채팅 삭제 기능
    delete:
      summary: 특정 사용자의 채팅 삭제
      operationId: delete_user_chat
      parameters:
        - name: user_id
          in: path
          required: true
          schema:
            type: string
          description: 사용자 ID
        - name: chat_id
          in: path
          required: true
          schema:
            type: string
          description: 채팅 세션 ID
      responses:
        "204":
          description: 채팅 삭제 성공 (No Content)
        "404":
          description: 사용자 또는 채팅을 찾을 수 없음
        "500":
          $ref: "#/components/responses/InternalServerError"

    # 채팅 내역 조회 (경로 변경)
    get:
      summary: 특정 채팅 내역 조회
      operationId: get_chat_history
      parameters:
        - name: user_id
          in: path
          required: true
          schema:
            type: string
          description: 사용자 ID
        - name: chat_id
          in: path
          required: true
          schema:
            type: string
          description: 채팅 세션 ID
      responses:
        "200":
          description: 성공적인 조회
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ChatResponse"
        "500":
          $ref: "#/components/responses/InternalServerError"

    # 메시지 전송 및 답변 반환 (경로 변경)
    post:
      summary: 메시지 전송 및 답변 반환
      operationId: chat_with_bot
      parameters:
        - name: user_id
          in: path
          required: true
          schema:
            type: string
          description: 사용자 ID
        - name: chat_id
          in: path
          required: true
          schema:
            type: string
          description: 채팅 세션 ID
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChatRequest"
      responses:
        "200":
          description: 성공적인 응답
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChatResponse"
        "500":
          $ref: "#/components/responses/InternalServerError"

# WebSocket 엔드포인트는 OpenAPI 3.0 표준에 직접 포함되지 않으므로 명세에 추가하지 않습니다.
# 대신 설명에 언급하거나 별도 문서를 참고하도록 안내할 수 있습니다.

components:
  schemas:
    ChatRequest:
      type: object
      properties:
        message:
          type: string
          description: 사용자 메시지
      required:
        - message

    ChatResponse:
      type: object
      properties:
        answer:
          type: string
          description: 챗봇 응답
      required:
        - answer

    ChatListItem:
      type: object
      properties:
        chat_id:
          type: string
          description: 채팅 세션 ID
        created_at:
          type: string
          format: date-time
          description: 채팅 생성 시간
        title:
          type: string
          description: 채팅 제목 (자동 생성 또는 사용자 지정)
      required:
        - chat_id
        - created_at

    CreateChatRequest:
      type: object
      properties:
        product_id:
          type: string
          description: 관련 제품 ID (선택 사항)
        title:
          type: string
          description: 채팅 제목 (선택 사항)

  responses:
    InternalServerError:
      description: 서버 내부 오류
      content:
        application/json:
          schema:
            type: object
            properties:
              detail:
                type: string
                example: 챗봇 서비스 오류 발생 